+++ src/test/java/roomescape/reservation/controller/ReservationApiControllerTest.java
@@ -15,7 +15,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import roomescape.config.ControllerConfig;
-import roomescape.payment.PaymentClient;
+import roomescape.payment.controller.PaymentClient;
 import roomescape.reservation.service.ReservationService;
 
 @WebMvcTest(ReservationApiController.class)
